fix: separate gui for darwin
